% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{copy.upper.lower}
\alias{copy.upper.lower}
\title{Copy upper triangle to lower triangle of a matrix}
\usage{
copy.upper.lower(m)
}
\arguments{
\item{m}{Square numeric matrix with upper triangle filled.}
}
\value{
Symmetric numeric matrix.
}
\description{
Fills the lower triangle of matrix \code{m} with the transpose of the
upper triangle, making the matrix symmetric.
}
\keyword{internal}
